dracut: add 99emergency-timeout
If Ignition fails, instead of immediately dropping to an emergency
shell, output a message indicating that Ignition failed and point
to ignition-validate. Offer to drop to an emergency shell if the
user presses Enter. If the user doesn't respond in 5 minutes,
reboot to try again. While waiting for the timeout, keep updating
the prompt with the time remaining until reboot.

Brought forward from the `99emergency-timeout` dracut module
present in Container Linux, with minor modifications:
https://github.com/coreos/bootengine/tree/c691205c0046e3828a8536d553aea7307ca3abee/dracut/99emergency-timeout

Closes: #35