Add ignition-virtio-dump-journal.service
Debugging failures in the initrd is annoying; this code
looks for a virtio-serial port named `com.coreos.ignition.journal`,
and runs as part of `emergency.target`.

I plan to change mantle to set up this port by default, so if
something fails in the initramfs we'll at least reliably get
the journal in a sane parsable format.

This is a special targeted subset of
https://github.com/coreos/ignition/issues/585

(cherry picked from commit 84c89f4a5fa5b32c39df01e6ec5fd081050daa62)