setup-user: Accept an existing /boot mount
For RHCOS, dracut mounts `/boot` in the FIPS case, and things
blow up if we try to mount it twice.

Also bigger picture at this point we have a crap ton of services
all trying to mount `/boot`; we might as well just add a `boot.mount`
unit to our initramfs.