Remove OnFailure= from targets, put in services
In systemd v245, which is in f32, the `OnFailure=` directive in target
units stopped working. This seems to be by design though:

https://github.com/systemd/systemd/issues/14142

If target units were never meant to handle this, this might be why we've
had issues like boot loops on failure in the past (see #115).

Anyway, let's just pepper our services with the directive instead.