systemd: coreos-autologin-generator: Add support for dropin to autologin as core user
Add support for autologin as core user when the coreos.autologin command line is set.
This generator overrides the ttyS0 getty service when the command line is provided and
does not persist the conf across reboots. Found this especially useful to debug networking
issues by getting a console without changing the ignition config.