99emergency-timeout/virtio-dump: wait before dumping
When systemd activates us as a result of switching to
`emergency.target`, journald might still be collecting and writing back
messages from various services. If we run too early, we'll be missing
out on those.

As a hack, wait until the 'Journal stopped' message from journald is
written before dumping the whole journal. Timeout at 30s.